ENH: Moves some code to DICOMPlugin base class
Make these methods available for us in other classes.
