+++ Modules/Scripted/DICOMPlugins/DICOMScalarVolumePlugin.py
@@ -171,15 +171,6 @@ class DICOMScalarVolumePluginClass(DICOMPlugin):
 
         return loadables
 
-    def cleanNodeName(self, value):
-        cleanValue = value
-        cleanValue = cleanValue.replace("|", "-")
-        cleanValue = cleanValue.replace("/", "-")
-        cleanValue = cleanValue.replace("\\", "-")
-        cleanValue = cleanValue.replace("*", "(star)")
-        cleanValue = cleanValue.replace("\\", "-")
-        return cleanValue
-
     def examineFiles(self, files):
         """ Returns a list of DICOMLoadable instances
         corresponding to ways of interpreting the
@@ -344,22 +335,6 @@ class DICOMScalarVolumePluginClass(DICOMPlugin):
                     loadable.confidence = .45
 
         return loadables
-
-    def seriesSorter(self, x, y):
-        """ returns -1, 0, 1 for sorting of strings like: "400: series description"
-        Works for DICOMLoadable or other objects with name attribute
-        """
-        if not (hasattr(x, 'name') and hasattr(y, 'name')):
-            return 0
-        xName = x.name
-        yName = y.name
-        try:
-            xNumber = int(xName[:xName.index(':')])
-            yNumber = int(yName[:yName.index(':')])
-        except ValueError:
-            return 0
-        cmp = xNumber - yNumber
-        return cmp
 
     #
     # different ways to load a set of dicom files:
